Merged changes from the 1.1 branch.
This is my first try at merging in Subversion.  Is it really this hard?
This is how I did it:

  svn merge -r 85842:86602 ../branches/1.1 .

Figuring out those revision numbers was complicated.

With Mercurial this would have been a breeze:

  hg pull ../branches/1.1

That command would simply do the right thing.  Isn't there an equivalent
for Subversion?  Hmph.