Include optimizer hints for Oracle.
With a big RelStorage database Oracle would otherwise choose an incorrect execution plan as it fails to notice that the pack_object table is not empty at this time. In the case of a large database (25 million rows in the OBJECT_STATE table) this query would fall flat and not return data for several days. The added hints make the optimizer choose the correct execution plan for this query whatever the size of either table.