PostgreSQL 8.1 compatibility: "drop table if exists" is not valid in 8.1.
This required changing "on commit delete rows" to "on commit drop", but
the "on commit delete rows" idea was not actually working anyway, since
"on commit delete rows" fires truncate triggers, not delete triggers.

Note that PostgreSQL 8.1 has reached end of life.  It might be wise
to immediately drop compatibility with PG 8.1.