Postgres tests work on Py3.
psycopg2 has a very nice strict handling of byte data on Py3.

An upside of this work is that we take advantage of the support for
BYTEA in the PG protocol and psycopg2 library and no longer base64
encode/decode all the binary data when talking to PG, which could come
with some performance benefits. It definitely reduces the code
duplication.