Use smaller buckets.
And tweak the bubbling strategy. This may wind up using slightly more
memory, but it should also be friendlier on eviction. The LRU policy is
also more nuanced, meaning the older a key is the more times it has to
be accessed to still survive the next GC.

When using 2 buckets (as before) the time for the included benchmark
changes from 2.69/0.36 for pop/read to 2.80/0.70. (Before any thread
changes, the numbers were 3.19 and 0.44, so we're still faster for
writing but slightly slower for writing.)