Implement ZEO cache tracing.
Fixes #106.

The ZEO tests pass, with a few numeric modifications due to the inexact
way we track invalidations and the different way we handle loads and
stores.

I think all the numbers look reasonable, but I want to eyeball them
somemore. Other eyeballs welcome! I also think we're incorporating all
the useful stats we can based on the table in #106 (since we don't
really track current/non-current).

I reduced the amount of duplicate tests down to a minimum; we expect the
ZEO analysis tools to work correctly and the duplicates just make it
harder to fix if we adjust our cache algorithm or tracing details.