Major optimizations for persistent cache files.
Typical operations for large caches will be 3x faster or more. Python 2
and <3.4 need more memory to write, though.

The benchmark is much better and more realistic which is what drove
these changes.

Introduced an option to not compress the local cache files; they aren't
compressed by default anymore because that was a big part of the
slowness.