Add a simulator to take the cache traces from UMass.
And run them, putting it in the docs.

The new code has a better hit rate in 12 of 15 examples, sometimes
a *much* better rate. The cases where it's behind, it's very close; the
biggest gap is 0.052 or 5%, and that's with a woefully undersized cache.

The new code is faster in all cases, sometimes by a lot.