Use a union to distinguish head/member nodes.
This gets rid of the confusing usage of 'frequency' to mean sum of
weights and len to mean count in one place, and weight in others.

Also rename 'len' to 'weight' to better reflect its purpose.

Somewhat surprisnigly, the extra attribute access required in python
didn't seem to have any performance impact on the benchmarks.