Better DBAPI compliance: don't depend on cursor.rowcount and full resultset buffering
This is a good thing in general.

In MySQL Connectors case, it means we can get away from using the
`buffered=True` parameter when we get a cursor, which is a necessary
step to being able to use prepared cursors (because the two types can't
be combined).

It also *massively* improves performance. Using a buffered cursor,
running the testmysql tests took 173s with the python driver; using the
default unbuffered cursor they go to ~129s, which is comprable to
PyMySQL's ~125s. The C implemenattion goes to 115s (I didn't get a
before number for it) which is comprable to mysqlclient's 111s.

It makes me wonder what using unbuffered queries in those two
implementations might look like. mysqlclient appears to use one API call
to buffer the entire result set (mysql_store_result). PyMySQL seems to
iterate "packet" by packet, while connector/python iterates "row" by
row, which appear to be the same. This requires further testing.