Checkpoint on sqlite-based cache and benchmarks.
Early results show that it can be a large win for memory usage, if we use a temporary database instead of in-memory (which has about the same performance, it seems).

More work needed to allow efficient use. It's still going to be slower at the micro level, it looks like, but the next step will be to try it at the macro level.