Simplify cache tracing and storing multiple states.
Move that into a new IStateCache wrapper. This frees up the
storage_cache implementation, and it removes any overhead when it's
not in use.

Storing multiple states after a transaction into the cache is now also
handled by its own separate logic we can test independently. This lets
the IStateCache operate at a higher level, and local caches no longer
pay the overhead of batching updates that we have to do for memcache.