Add more aggressive validation and, when possible, corrections for certain types of cache consistency errors.
Previously an ``AssertionError`` would be raised with the message
"Detected an inconsistency between RelStorage and the database...". We
now proactively try harder to avoid that situation based on some
educated guesses about when it could happen, and should it still
happen we now reset the cache and raise a type of ``TransientError``
allowing the application to retry. A few instances where previously
incorrect data could be cached may now raise such a
``TransientError``.

I also noticed that the poller's "list_changes" method does make
certain guarantees, and so we started to rely on them. This helps us
detect any issues initially building up the local delta maps.