Validate older cache entries when loading from a persistent cache.
This makes the cache useful again for objects that rarely change in write-heavy databases. The cost appears to be less than 1s per 100K cached objects.

This does mean that we should more carefully track frequency of hits and evict old objects that are never actually used, though.

Also, we should propagate invalidations down to the persistent cache layer, but this isn't done here.