Fix #270, fix #271 and fix #272
(Yes, all in one commit. I'm sorry.)

Introduce a strict mode for testing, and use the server-side cursor in
MySQLdb when that's the case. It raises the ProgrammingError we were
seeing in #270 consistently, so find and fix all of those.

At the same time, make the gevent MySQLdb driver more cooperative, and
make OID GC block for less time.