Central tracking of polling.
Mostly seems to work. There's something broken with persistence and undo.

Much more on centralized polling.

Everything is working locally. Only one connection ever does a poll at any one time, the rest pick it up when it gets ready.

Needs more unittesting and performance tuning.

Also saving can take advantage of this information in a big way, so that needs done.

This gets us part way to a smarter local cache.

Fix unit tests

Need to consume the cursor even if we're not going to produce changes.

mysql-connector-python requires this.

Debugging for appveyor

In the old procs used on appveyor, the temp table data was leaking across aborts because a mysql abort doesn't fully abort.

In the old procs used on appveyor, the temp table data was leaking across aborts because a mysql abort doesn't fully abort.