Fix several race conditions when packing history-free storages
Fixes #325.

Use a combination of changes in ordering and more strongly
consistent (``READ ONLY REPEATABLE READ``) transactions.

The longer running load connection is not expected to be a problem;
we've run a test with an active load connection loading objects for 8
hours while the database was being written to without any problems in
a 60 million object database.

Reported in :issue:`325` by krissik with initial PR by Andreas
Gabriel.