Docs: Remove exclusion of PostgreSQL 12. We do support it.
The query I was worried about does not get inlined because it has side-effects. I think that changed from the first descriptions I read.

All tests pass. We know that they test for that query not working because several tests (checkTL_*) fail if we manually remove the locking from that query.

Fixes #364.