PostgreSQL: Send the application_name at connect time.
This is faster and more importantly works even when connecting to a hot standby. SET SESSION application_name does not work, eventhough it is documented that SET does work.

Also handle a port specification in pg8000.

Fixes #378 and further refines #376 (this time I set up a hot standby and ran zodbshootout's cold benchmark against it so that much at least works).