Fix sqlite+gevent deadlocks.
The problem happened if we switched away from a connection holding locks and into one that wanted to take those same locks *without* switching. This could be non-deterministic.

The solution was to move those operations that take locks into gevent's threadpool so they're guaranteed not to block the event loop. Fixes #382.

Add tests for this scenario. Also, CI wasn't properly testing gevent monkey-patching because of a bug in gevent.monkey. Workaround that and fix the issue that the gevent.monkey bug was hiding (around mixing of real native threads and greenlets).