Fix #385 by eliminating the multiple result sets for lock_objects_and_detect_conflicts
It turns out the root of the problem was that there's no way to
reliably know whether we need to, or even can, go around the event
loop to get the second and subsequent result sets. And previously, it
could either be the first result set or the second result set that
blocked taking locks.

Now there's only one result set, so our normal way of waiting for IO
to arrive works.

Also throws in some minor optimizations targeted to avoiding extra
trips around the event loop.