Initial implementation of specialized copyTransactionsFrom for HF storage.
Increases the copy rate from 1.01MB/s to 5.58MB/s on a sample of transactions, or 180000 objects to 417000 objects.

Outstanding issues:

zodbconvert's incremental mode doesnt work.
zc.zlibstorage doesn't work on the destination during the copy. Maybe that's ok.
Because you could just not use a decompressing wrapper on the source to get identical copies.
Logging is messed up.