Using a Bucket instead of a BTree can be nearly 50% faster on MVCC _vacuum.
This is because the major bottleneck is the keys() method, which is a BTree multiunion. That function is at its fastest given a Bucket or Set.

I'm not sure what unintended impacts this change might have. Probably none? Bucket has very different performance charecteristics as it gets larger, but ZODB's fsIndex claims that's not an issue with a 65K entry bucket (at least as long as data is inserted ordered --- I couldn't detect a difference with buckets of size 250). Most buckets we use are probably on the smaller end of that scale, except the initial one we restore from cache.

I'm still contemplating other data structure or algorithm improvements.