Move the core of the loop into Cython.
This gets us to an incredible 3 seconds on the benchmark for the small list of OIDS (compared to 44s)

It gets us to 34s on the big list of OIDs (compared to 55s originally, or 29s with BTree Buckets; but there the small list still took 22s).

The next step might be to try maintaining an aux structure containing the newest OID/TID map consolodated.